[qa] move the verify_file_callback in utils.py to reuse this in other tests units
